(hel-741) Extraction du Type « SelectedRows » dans le composant commun
